% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dlParadigmBin}
\alias{dlParadigmBin}
\title{Download PARADIGM binary and return its location}
\usage{
dlParadigmBin()
}
\value{
full path of downloaded PARADIGM binary
}
\description{
Download PARADIGM binary and return its location
}
