% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltNeiStt.R
\name{pltNeiStt}
\alias{pltNeiStt}
\title{Plot a heatmap of pathway and omic states of a protein and its
pathway neighbors}
\usage{
pltNeiStt(cn_state_mat, rna_state_mat, fltmat, fpth, protein)
}
\arguments{
\item{cn_state_mat}{A matrix of PARADIGM CNA states with rows as genes and
columns as samples. It is the same matrix as the output
from \code{ppCnInp()}.}

\item{rna_state_mat}{A matrix of PARADIGM RNA states with rows as genes and
columns as samples. It is the same matrix as the output
from \code{ppRnaInp()}.}

\item{fltmat}{A matrix contains filterd IPL with rows
as 'entity' and column as samples. This is the output from
\code{fltByPerm()}.}

\item{fpth}{Name of a pathway file for PARADIGM.}

\item{protein}{Name of the protein to plot. It requires to have CN and RNA
state data, as well as pathway data from the input.}
}
\value{
A heatmap of pathway and omic states of a protein and its
pathway neighbors
}
\description{
Plot a heatmap of pathway and omic states of a protein and its
pathway neighbors
}
\examples{

fpth = system.file('extdata/Pth/tiny_pth.txt', package='MPAC')

fcn  = system.file('extdata/pltNeiStt/inp_focal.rds',       package='MPAC')
frna = system.file('extdata/pltNeiStt/inp_log10fpkmP1.rds', package='MPAC')
fflt = system.file('extdata/pltNeiStt/fltmat.rds',          package='MPAC')

cn_state_mat  = readRDS(fcn)
rna_state_mat = readRDS(frna)
fltmat = readRDS(fflt)
protein = 'CD86'

pltNeiStt(cn_state_mat, rna_state_mat, fltmat, fpth, protein)

}
