% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppInp.R
\name{ppCnInp}
\alias{ppCnInp}
\title{Prepare input copy-number (CN) alteration data to run PARADIGM}
\usage{
ppCnInp(cn_tumor_mat)
}
\arguments{
\item{cn_tumor_mat}{A matrix of tumor CN focal data with rows as genes
and columns as samples}
}
\value{
A matrix of CN state for PARADIGM
}
\description{
Prepare input copy-number (CN) alteration data to run PARADIGM
}
\examples{

fcn = system.file('extdata/TcgaInp/focal_tumor.rds', package='MPAC')
cn_tumor_mat = readRDS(fcn)

ppCnInp(cn_tumor_mat)

}
