% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppInp.R
\name{ppRealInp}
\alias{ppRealInp}
\title{Prepare input copy-number (CN) alteration and RNA data to run PARADIGM}
\usage{
ppRealInp(cn_tumor_mat, rna_tumor_mat, rna_normal_mat, threads = 1)
}
\arguments{
\item{cn_tumor_mat}{A matrix of tumor CN focal data with rows as genes
and columns as samples. A value of 0 means normal CN,
> 0 means amplification, and < 0 means deletion.}

\item{rna_tumor_mat}{A matrix of RNA data from tumor samples with rows as
genes and columns as samples}

\item{rna_normal_mat}{A matrix of RNA data from normal samples with rows
as genes and columns as samples}

\item{threads}{Number of threads to run in parallel. Default: 1}
}
\value{
A SummarizedExperiment object of CN and RNA state for PARADIGM
}
\description{
Prepare input copy-number (CN) alteration and RNA data to run PARADIGM
}
\examples{

fcn = system.file('extdata/TcgaInp/focal_tumor.rds', package='MPAC')
ftumor = system.file('extdata/TcgaInp/log10fpkmP1_tumor.rds', package='MPAC')
fnorm = system.file('extdata/TcgaInp/log10fpkmP1_normal.rds', package='MPAC')

cn_tumor_mat = readRDS(fcn)
rna_tumor_mat = readRDS(ftumor)
rna_norm_mat  = readRDS(fnorm)

ppRealInp(cn_tumor_mat, rna_tumor_mat, rna_norm_mat)

}
