% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPrd.R
\name{ppRunPrd}
\alias{ppRunPrd}
\title{Prepare required files to run PARADIGM}
\usage{
ppRunPrd(pat, cnmat, rnamat, outdir, file_tag=NULL)
}
\arguments{
\item{pat}{Sample ID}

\item{cnmat}{CN matrix}

\item{rnamat}{RNA matrix}

\item{outdir}{Output folder to save all results.}

\item{file_tag}{A string of output file name tag. Default: NULL}
}
\value{
None
}
\description{
Prepare required files to run PARADIGM
}
