% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clSamp.R
\name{clSamp}
\alias{clSamp}
\title{Cluster samples by pathway over-representation}
\usage{
clSamp(ovrmat, n_neighbors = 10, threads = 1)
}
\arguments{
\item{ovrmat}{A matrix of gene set over-representation adjusted p-values
with rows as gene sets and columns as samples. It is the
output from \code{ovrGMT()}.}

\item{n_neighbors}{Number of neighbors for clustering. A larger number is
recommended if the size of samples is large. Default: 10.}

\item{threads}{Number of threads to run in parallel. Default: 1}
}
\value{
A matrix of sample clustering result with rows as samples and a
column of cluster index
}
\description{
Cluster samples by pathway over-representation
}
\examples{

fovr = system.file('extdata/clSamp/ovrmat.rds', package='MPAC')
ovrmat = readRDS(fovr)

clSamp(ovrmat)

}
